Move tifffile to misc/borrowed
