+++ hyperspy/misc/borrowed/tifffile.py
